[autotest] Prevent firmware updates if there's a hardware mismatch.
From time to time, DUTs are found to have the wrong board label.
In the presence of automated firmware update, that could result in
a DUT trying to update to firmware that isn't appropriate to the
DUT's hardware.

This adds a sanity check to the firmware update verifier that
prevents an update from installing firmware for one kind of hardware
if the DUT is running firmware for a different kind of hardware.

BUG=chromium:611242
TEST=run this in a local instance on a DUT needing upgrade

Change-Id: I0e16d26fad0e8eb0a5e74f9204ed7024088a2212
Reviewed-on: https://chromium-review.googlesource.com/412383
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>