bluetooth_device_xmlrpc_server: use same mainloop
The mainloop could be reused without the need of creating
a new one each time.

This patch also fixes the return value of advertising_async_method().
It returns an empty string on success, None if the _advertising
interface manager is None, and an error string if the dbus method
fails. This is useful to verify the dbus method error in subsequent
patches.

BUG=chromium:602461
TEST=None. Will be tested together with the CL:
bluetooth_AdapterLEAdvertising: add 16 test cases for multi-advertising

Change-Id: I0aeb2a8dc7e9e0d6bebca7898f7b745afbdece86
Reviewed-on: https://chromium-review.googlesource.com/412908
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>