cros/bluetooth: unregister advertisement
This patch provides the method to unregister an advertisement
and its corresponding test.

Note that to unregister an advertisement, it is required to use
the same self._advertising interface manager. This is because
bluez only allows the same sender to invoke the
UnregisterAdvertisement dbus method.

BUG=chromium:602461
TEST=None. Will be tested together with the CL:
bluetooth_AdapterLEAdvertising: add 16 test cases for multi-advertising

Change-Id: I0ba76d72e07e9abf319f6fd118469e7c285e0db0
Reviewed-on: https://chromium-review.googlesource.com/412909
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>