abstract_ssh: get_file: retry rsync on possible network errors
We see a number of rsync failures in the builders which may be
due to transient network errors and may be recoverable with a
single retry, so do that.  The cost of the retry is expected
to be small, and the payback could be large.

BUG=chromium:664587
TEST=none

Change-Id: Ifbcf4e49c45befa6d1ab66ba860df60adbb17637
Reviewed-on: https://chromium-review.googlesource.com/413330
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>