Add security_RunOci
The security_RunOci test is formatted similarly to security_Minijail0.
There is a main parser that runs tests specified as text files.  These
files specify the program to be run and the expected output.  This
program is run in a container with a remounted rootfs as its root.
Anything available in a normal Chromium OS build can be run from the
test.

BUG=665622
TEST=This is a test (security_RunOci)

Change-Id: I61b9548a2147985ca8cea1ea3f8566c0e8934cac
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/413154
Reviewed-by: Mike Frysinger <vapier@chromium.org>