[autotest] audio_quality_measurement: Ignore false alarm
Ignore false alarm of "noise before/after sine wave" when it is too
close to the start/end of the sine wave.
We determine start/end of sine wave by frequency of a block.
We detect noise before/after sine wave by magnitude of a block.
The block near start/end of sine wave will have magnitude that
might be incorrectly treated as noise. Add a 0.01 duration to account
for that edge effect.
This is especially easy to seen in WebRTC loopback test because the
magnitude of sine wave is large in the beginning, and then it becomes
much smaller when it stablizes.

Also add debug message for start/end detection.

BUG=chromium:667674
TEST=run audio_AudioWebRTCLoopback test.

Change-Id: Ibd9bc95c896b290e83e7651093db121d2388b6f1
Reviewed-on: https://chromium-review.googlesource.com/414966
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>