security_RunOci: Add Oci tests
Extra command line arguments can be specified and should be appended to
those provided in config.json.  This test checks that an argument gets
to the specified program.

Add a test for alt-syscall.  This requried being able to check stderr
from the run process.

The containers will be run from crosh which runs as chronos.  To better
test the actual way the code will run, run the test containers as
chronos.

BUG=none
TEST=security_RunOci

Change-Id: I0e26549789af021eb23219738121cb84d77eaab7
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/415525
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>