mash: Add desktopui_MashLogin autotest
This tests that chrome --mash starts up, reaches the first OOBE screen,
and can log in. The purpose is to catch startup crashes specific to mash
and the Ozone DRM graphics backend (which is only used on real hardware).

This test starts on the bvt-perbuild suite. I will move it to bvt-cq once
we prove it is stable.

BUG=chromium:669243
TEST=desktopui_MashLogin

Change-Id: I4c2a2035d5a3fc92b12556ade35d536b5671fe93
Reviewed-on: https://chromium-review.googlesource.com/415267
Commit-Ready: James Cook <jamescook@chromium.org>
Tested-by: James Cook <jamescook@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>