[autotest] Ignore missing "chrome_version" in stable version assignment
In the Omaha JSON file, some entries (notably, the 'storm' product
family) don't have a "chrome_version" setting.  Beta channel entries
like this were causing the assign_stable_images script to fail.

The only board that currently causes failure is 'gale', which isn't
in the lab.  So, as a stopgap, we're changing to just ignore these
entries.

Eventually, this problem will show up in 'whirlwind', at which point
we'll lose the ability to assign a stable version for that board,
until there's some better fix.  That's expected to happen when R57
hits Beta.

BUG=chromium:669944
TEST=unit tests, run with --dry-run

Change-Id: If56ffccc2f6e4a66fbdfacb110431301d5a6736f
Reviewed-on: https://chromium-review.googlesource.com/415566
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>