Fixed bug with folders being sent over rsync
When given a folder rsync will always place the folder inside
the destination folder. This is a problem when the destination
folder is the same name as the source folder as you will end up
with folder_name/folder_name. Sending files now checks if the
source and destination have the same base name and if so it is
ommited from the destination.

BUG=b:33077320
TEST=Ran simple acts test.

Change-Id: I2ed676145a1c3ff80984b699086f2cafa73f95c8
Reviewed-on: https://chromium-review.googlesource.com/415568
Commit-Ready: Benjamin Peake <bpeake@chromium.org>
Tested-by: Benjamin Peake <bpeake@chromium.org>
Reviewed-by: Tom Turney <tturney@google.com>