Pulled out container code from testing code in acts_lib
The ActsContainer required a testbed to be passed in
because the testing components needed it to run.
However there are conditions where acts needs
to be installed in a container but not run for tests.
This code has been pulled out into its own class to
allow using ACTs for more than just running tests.

BUG=b:33302700
TEST=Ran simple acts test.

Change-Id: I75ed5c80125b7871d771429386c0a1e0b2418394
Reviewed-on: https://chromium-review.googlesource.com/416300
Commit-Ready: Benjamin Peake <bpeake@chromium.org>
Tested-by: Benjamin Peake <bpeake@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>