Android apks now installed from any custom artifact
Previously apks could only be installed from the preloaded
artifacts. This meant that tests.zip had to be a part of
provisioning and any apks outside of that could not be
installed. The adb_host code has been changed to not download
tests.zip until installing sl4a, and to allow additional
apks to come form any artifact. By default if one is
not given it comes form tests.zip.

BUG=b:33214713
TEST=Ran android_EasySetup

Change-Id: I751ed042189509dad156830903f7c7e8ca1de8b8
Reviewed-on: https://chromium-review.googlesource.com/417193
Commit-Ready: Benjamin Peake <bpeake@chromium.org>
Tested-by: Benjamin Peake <bpeake@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>