FAFT: move bypass_rec_mode out of mode_aware_reboot
There is some code in mode_aware_reboot that takes
care of handling the recovery FW screen.  However,
it is impossible to tell the difference between the
DUT being in the FW screen and the DUT being off.
As a result, this functionality also will unintentionally
recover a DUT when it is actually offline instead of
letting the test fail with a ConnectionError.  As a
result, we get false passes (ie: when ECLidSwitch doesn't
come back online because of a bad lid switch, FAFT
will actually recover the device and the test will pass).
Pulling the bypass_rec_mode function out of
mode_aware_reboot and fixing tests so that they
explicity call bypass_rec_mode.

BUG=chrome-os-partner:60721,chrome-os-partner:60208
BRANCH=None
TEST=Ran faft_bios, faft_ec suites on kevin

Change-Id: I9946752944311a23a62fdac319d91f0d87b7f835
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/418014
Reviewed-by: Wai-Hong Tam <waihong@google.com>