update the snaplen to cover AMSDU aggregation cases
While running this test case with AMSDU aggregating wifi card,
the test has been failing as the pcap parser was not able to detect
the AMSDU subframes from the chopped frame. To detect the UDP stream
and the src ip addresses we need to collect the full container. The
shanpshot length (snaplen) tcpdump parameter set to 400 was not enough
to decode the A-MSDU frame.

By spec the aggregated frames in 11n mode can by up to 7935bytes.
In case of the NFA344A module, this was around 3560bytes, but lets use
the max spec value just in case.

BUG=chrome-os-partner:50854
TEST=on Skylake and Apollo Lake platforms with Qualcomm's NFA344A NIC
Change-Id: I8036647eb367912c066c951b6e13b6e997f0bdd1
Signed-off-by: Bartosz Markowski <bartosz.markowski@tieto.com>
Reviewed-on: https://chromium-review.googlesource.com/417431
Commit-Ready: Waldemar  Rymarkiewicz <ext.waldemar.rymarkiewicz@tieto.com>
Reviewed-by: Waldemar Rymarkiewicz <waldemar.rymarkiewicz@gmail.com>
Reviewed-by: Sameer Nanda <snanda@chromium.org>