crash: Generalize crash testing library
In a previous CL, the reusable logic in logging_UserCrash was extracted
into library code. This CL adds optional parameters for customizing that
logic when the crasher process is running in a container.

For example, the test requires the PID of the crasher process. The
crasher may be running under su, so it prints its PID and the test
parses it. However, when the crasher is running in a different PID
namespace, the printed PID is incorrect. With this CL, the crasher
can be instructed to send its PID over a Unix domain socket, which
converts it to the receiver's PID namespace. The testing logic can
also be customized with a closure to replace the default crasher
invocation, which will be used by the cheets_NativeCrash test to
receive the PID before the crash.

BUG=b:27503472
TEST=cheets_NativeCrash
TEST=logging_UserCrash

Change-Id: Ia44bf0f5cef8fd54fb2d71ea20c42fbf278e27e0
Reviewed-on: https://chromium-review.googlesource.com/416358
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>