perf_uploader: Append platform_name with '.arc'
Since ARC++ is introduced on CrOS, we duplicated several control files
(with suffix of '.arc') to measure if there is any performance impact
between ARC++ is enabled/disabled for performance benchmarks. However,
the chromeperf which stores test results could not distinguish the
results as all key/bot names are the same for a test. Therefore, the
patch addresses this issue by appending the '.arc' as the platform_name
which is used for bot name in chromeperf if ARC++ is enabled for a test.

BUG=chromium:669717
TEST=None

Change-Id: Id7dfa7a5eb86116916eda1e114fd750e7b7ab311
Signed-off-by: Chung-yih Wang <cywang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/418595
Reviewed-by: Hsu-Cheng Tsai <hctsai@chromium.org>