autotest: fix that "beaglebone image not updated"
Beaglebone image is not well updated to stable version. This CL fixes this
issue by choosing a devserver in restricted subnet.

BUG=chromium:657662
TEST=Modify files on hot, and test it in chromeos1-row1-rack11-host1, update
its servo from 8638.0.0 to stable version 8866.0.0.
Verify it locally with the same settings as hot, successfully resolve a
devserver in the same subnet with a servo hostname.

Change-Id: I1ac224cdab0bec35dbc27234c1ecb4658691b545
Reviewed-on: https://chromium-review.googlesource.com/418400
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>