SSHHost.run(): add API to retry ssh calls on probable ssh failure
We see a very small rate of probable ssh failures (timeout,
or command status 255), which usually translate into complete
failure for the testing in progress.  Since a single run of
the release builders involves millions of ssh commands, even
a small failure rate results in frequent breakage.

The failures are not understood, but because the ssh commands
following a failure usually succeed, it is likely that a simple
retry strategy, when possible, will recover from them in most
cases.

Each failure is either reported as ssh error 255, or by a timeout
enforced by the python code running the ssh command.  In either
case we cannot tell whether the failure is due to ssh and/or
network errors, or problems with the command executing on the
DUT.  So we just add the keyword parameter "ssh_failure_retry_ok"
to SSHHost.run(), so that individual calling sites can choose
the new behavior, with the understanding that the retries will
happen on "probable" ssh errors.

Failure modes of ssh are complicated by the presence of the
ssh control master, which sits in between the ssh client and
the ssh daemon on the DUT.  So during the retry attempts, we
eventually restart the control master.

This is the flow for calls that specify
sh_failure_retry_ok = True:

1. try first time, but disable exceptions on timeout
2. if success or DNS error, same behavior as before (return on
success, retry once on DNS error)
3. if timeout or status 255 is returned, retry identically
4. if timeout or status 255, retry again after restarting the control
master and resetting the timeout behavior to the original one.

When switching a call site to use ssh_failure_retry_ok, consider
that the worst-case timeout could be 3x longer than the specified
timeout, and adjust it as needed.  This should be necessary
only 1. for very long timeouts or 2. for code that can recover
from such timeouts (if it's going to fail anyway, it doesn't
matter if it takes a little longer).

BUG=chromium:664587
TEST=none

Change-Id: I1f61cdba98b6ed1f3543e5ab38fa7f5bfc37bdc3
Reviewed-on: https://chromium-review.googlesource.com/418691
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>