security_ASLR: Make finding processes more robust
There have been a few flaky runs where the first line of `pidof ${name}`
is not the process we were looking for and goes away before we are able
to look for /proc/${pid}/maps. This change removes the code path where
we only look at `pidof` and instead always requires that the parent
process is specified to reduce flake.

BUG=chromium:618382
TEST=test_that cyan security_ASLR
TEST=test_that lakitu security_ASLR

Change-Id: I5d296657da7c6561da86266724692f9659c15163
Reviewed-on: https://chromium-review.googlesource.com/419096
Commit-Ready: Luis Hector Chavez <lhchavez@chromium.org>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Andrey Ulanov <andreyu@google.com>