[autotest] Remove temp var n_scheduled
Get rid of the premature optimization of storing the length of a
list.  This optimization has zero performance impact, but it
introduces an extra local variable.

NOTE that the new behavior is not strictly identical to the original
behavior, in that if an exception is raised after partially scheduling
tests, we would previously return 0 tests scheduled, while we would
now return the number of tests scheduled.

1. I dont think the return value is actually used anywhere.
2. The original commit that added the except clause doesnt seem to be
   intending this specific behavior.
3. The new behavior is semantically more correct, since those
   scheduled tests have in fact been successfully scheduled.

BUG=None
TEST=None

Change-Id: I95defdc0d5d0be5e6ac154e2dada3898b3cfa9c8
Reviewed-on: https://chromium-review.googlesource.com/419107
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>