autotest: do not rebase before running automated_deploy
The instructions on our internal website already tell you to make sure
you are at ToT before running automated deploy.

The `git pull --rebase` command failed for me (because I was not on a
tracking branch) and caused the automated_deploy script to fail.

Either we should make the script not crash when that fails, or we can
just get rid of that git pull.

BUG=None
TEST=Ran automated_deploy locally to do a deployment.

Change-Id: Ief7b584989cc456fc674dd06ff8fb388f8b91758
Reviewed-on: https://chromium-review.googlesource.com/419176
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>