FAFT: Remove asynchronous reboot/wake calls..
The httplib that the XMLRPC module uses, does not seem to be threadsafe.
In particular, it was somewhat common for firmware_ECWakeSource to fail
with a ConnectionError raised by the httplib.  This was because the wake
methods used by this test were being called asynchronously and were
causing collisions for the same HTTPConnection object.  This commit
changes the tests to use synchronous calls instead.

BUG=chrome-os-partner:59753
BRANCH=None
TEST=Run firmware_ECWakeSource on sentry and verify it succeeds.
TEST=Run firmware_DevModeStress on sentry and verify it succeds.

Change-Id: I9c9abfd0bcec9e3c93a9068bc6c5b1af5fdf6274
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/419067
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>