[autotest] Distinguish "get_version" from "get_image_name"
For CrOS builds, the AFE class for access to the get_stable_versions
RPC functions was formatting the stable version as a complete image
name, not just a version number.  This was a problem for two reasons:
  * Some callers needed the version number, not the formatted image
    name.
  * The get_version() method was returning values inconsistent with
    the behavior of get_all_versions() and set_version(), both of
    which use version numbers.

This changes the get_version() method to return a version string,
and adds a new get_image_name() method for returning the image name.
It also converts callers that need an image name to use the new
method.

BUG=chromium:673836
TEST=Manually invoke the new method from python CLI

Change-Id: I7b3a3e6ae10438b00c70e93e29ca250c05cfd2a4
Reviewed-on: https://chromium-review.googlesource.com/419329
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>