autotest: log auto-update logs path after logging the error traceback
Sheriffs have recently reported that they are not aware of the log for a
failed autoupdateprocess, which can be very useful for debugging. This CL
adds the path of the autoupdate logs after the error traceback so that
sheriff can go to that directory to find detailed auto-update logs.

BUG=chromium:673552
TEST=Test it locally. Manually set devserver's return and then call
ds.auto_update() to see whether this logging is successfully printed.

Change-Id: I0f023514682c015514ed0a43c8cb0f4807010346
Reviewed-on: https://chromium-review.googlesource.com/419901
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>