[autotest] Remove old priority hack
This is from 2013, Im assuming the priorities have all been migrated
over.

The reason for doing this is that the sole purpose of this function is
to remove invalid keywords from passing locals() to create_job_common,
so this is a step toward removing locals() and this function.

BUG=None
TEST=CQ

Change-Id: Ia5b527b509863bbe12ff047628c23a988dbc32c9
Reviewed-on: https://chromium-review.googlesource.com/420178
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>