[Autotest] Parse stylus hw_id and fw_id.
Wacom styluses don't actually have a hw_id the same way that our
existing touchpads and touchscreens do.  Use 'wacom' as a workaround
for our autotests.

The updater output also differs, so modify the pass_terms in the
UpdateErrors test to look for the correct terms in the wacom case.

Add a control file to check for stylus update.

TEST=ran on Kevin and upcoming stylus devices.  Ran on non-stylus device.
BUG=chromium:57999

Change-Id: Iff0e2cd12c71969e3f7b49799ff029eae3c24bd1
Reviewed-on: https://chromium-review.googlesource.com/421223
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>