[autotest] Fix dut_status_unittest time comparisons.
The logic for unit testing `dut-status --until` was wrong.  It was
supposed to test the "--until" value was "now" in the sense that
it was in-between two timestamps that bracketed the calculation of
the value.  One of the comparisons was for "<=" instead of ">=",
leading to occasional flake.

This changes to use a mock to force `time.time()` to return a
pre-computed value.

This change also contains other miscellaneous (non-functional)
clean-up of the unit tests.

BUG=chromium:675667
TEST=run the unit tests using a patch to provoke the bug.

Change-Id: I3b3a6fc10d7715cf83103f67e272ccc3a216efb0
Reviewed-on: https://chromium-review.googlesource.com/422427
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>