Setup wizard skip no longer fatal
The setup wizard on some builds does not exist. When trying to skip
the setup wizard on these systems the code ends up crashing. Failing
to skip the setup wizard does not need to be fatal so it will now
log an error and continue on with provisioning.

BUG=b:33424242
TEST=Provisioned phone with no setup wizard.

Change-Id: I0fa2df11e1049835ea66f12f4847c61ff5662062
Reviewed-on: https://chromium-review.googlesource.com/422411
Tested-by: Benjamin Peake <bpeake@chromium.org>
Tested-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Commit-Queue: Bindu Mahadev <bmahadev@chromium.org>
Trybot-Ready: Bindu Mahadev <bmahadev@chromium.org>