abstract_ssh: extend get_file API to ignore potentially dangling symlinks
In collect_logs(), we copy symlinks that end up dangling in the target.
A later copy results in errors and test abort, as explained in the bug.
This extends the API to explicitly ignore symlinks that can potentially
dangle in the target.  We could argue that this should be the standard
behavior, but we prefer not to change it.

BUG=chromium:467066
TEST=None

Change-Id: I930f91b66d2cd8b5f615b001962b5b3fc095b4d2
Reviewed-on: https://chromium-review.googlesource.com/422431
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>