scheduler: Don't schedule host_jobs when inline acquisition is false.
host_scheduler, when run as an independent service assumes that it is
the sole service scheduilng host_jobs. A fallback path in scheduler
would try to schedule these jobs. This situation was never possible in
the normal case, but the fallback was dangerous -- it could leave the DB
corrupted.
Instead, report errors in the fallback path and refuse to schedule the
host jobs.

BUG=chromium:675048
TEST=unittests, local jobs.

Change-Id: I60bb56ded41c34484f6e281fdd12ace5601e739d
Reviewed-on: https://chromium-review.googlesource.com/422535
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>