[Autotest] Enterprise: add parameter to skip policy verification.
Some policies have the key "default_for_enterprise_users".  These
policies have a mandatory value when the policy is not set.

E.g. when AllowDinosaurEasterEgg is not set, the value on
chrome://policy page is False.  Before, the setup_case function
expected that the policy_value matches chrome://policy.

Add a parameter to setup_case to allow us to skip verification in
this case - it will be up the individual test to check for whatever
value is expected.

TEST=Ran it
BUG=None

Change-Id: I8f808b2df9145c39f37437c03d4090f844938741
Reviewed-on: https://chromium-review.googlesource.com/422915
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>