[autotest] Use provisioning in place of parameterized jobs.
Formerly, when the create_job() RPC was passed an image for testing,
the handling was to create a special parameterized job that would
invoke autoserv with the `--image` option, so that the build would
be installed directly.  However, the specified `image` parameter
couldn't be passed to shards, so the feature could only work for
DUTs assigned to the autotest master.

This changes handling such that when an image is specified at job
creation time, the image is converted to a `cros-version` label
dependency.  This allows the specified build information to be
passed to shards.  Additionally, after this change, the code
relating to parameterized jobs becomes dead (along with the related
image specification/installation code), enabling additional code
cleanup.

BUG=chromium:628694
TEST=create jobs with images via web UI and CLI in a local instance

Change-Id: Id1fb5444698bb776ca5367d00e74ad3dbbd1a88c
Reviewed-on: https://chromium-review.googlesource.com/423389
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>