[autotest] Stablize external microphone test
Previously, we only rely on second_peak_ratio to check DC component when
it happens to be the second largest component.
Now we set different threshold for DC component and always check it to
be small enough.

Ignore 50 and 60 Hz for external microphone test.

BUG=chromium:661722
TEST=run the test on chell
Change-Id: Idc7527992b905b9d41b25d63dab91ad54aa2ec91
Reviewed-on: https://chromium-review.googlesource.com/424242
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>