security_RunOci: Don't list devices, list namespaces
Listing devices doesn't work because there isn't a way to create them
after mounting /dev when starting a container with minijail.

Do list namespaces that the container should enter, the container code
will soon check these so that it isn't forced to enter all namespaces if
it doesn't make sense.

BUG=none
TEST=security_RunOci

Change-Id: I2a73dfd689dc693fa97a876b3821640feb0a5a2b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/424188
Reviewed-by: Mike Frysinger <vapier@chromium.org>