ssh_ping: retry on ssh failure
This makes ssh_ping not give up on the first ssh error.  Currently
we see cases in which the "ssh true" command hangs for unknown reasons
(DUT is up); no retries are made; the test fails, and the entire build
fails.

We also need to accept this keyword parameter in servo_host::run(),
although this problem is not known to exist for servo_host, and thus
we ignore its value for now.

BUG=chromium:671760
TEST=none

Change-Id: Ie2a396a7c8b06f461f57fe823dd8664372806d16
Reviewed-on: https://chromium-review.googlesource.com/425779
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>