Rewrite autoupdate_EndToEndTest to use the Devserver-on-DUT provision flow.
The test will now follow the same flow as the lab provisioning and cros
flash. This is done with the devserver cros_au RPC. The rootfs
and stateful payloads will be copied to the DUT along
with the devserver code. It will then startup a devserver on the DUT and check
for an update at 127.0.0.1 eliminating network issues.

This test specifies the exact Google Storage URIs for the
payloads that we want to use (source and target). I have done the code for
having cros_au handle exact payload filenames in previous CLs.

The basic flow of this test will now be:
-Stage payloads on a devserver
-Install the source image on the DUT using cros_au RPC
(via dev_server.auto_update()).
-Call cros_au (again via dev_server.auto_update) for the target update and
parse the hostlog collected during the update AFTER the update is complete.
Before this change we made repeated requests to a hostlog URI during the update.

In this CL I have done these things:

1) Remove the OmahaDevserver code from the test. This was used to kick off
a devserver instance for the test on a remote machine. This is no longer
needed.

2) The test verifies an update is succesful by parsing a event "hostlog"
during an update. We will still be parsing this hostlog to check for all
of the expected steps BUT it will now be done with a log file after the
update instead of making many requests to a remote URL during the
update. (The setup for returning the hostlog was done in previous CLs)

3) The TestPlatform interface no longer made sense as is and since there is
now only one implementation (ChromiumOSTestPlatform) I have updated the
API where applicable so that it fits the new logic.

4) Extract the correct values from a Google Storage URI (build_name and
payload_filename) and pass these to cros_au (via dev_server.auto_update).
cros_au is now used to update to the source and target payloads.

5) Stateful update is controlled by cros_au too. The code for selecting
the stateful_update_bin is done by chromite/lib/auto_updater.py

6) Update the control file to have up to steps for running locally.

BUG=chromium:709710
BUG=chromium:689105
TEST=test_that <ip> autoupdate_EndToEndTest locally
TEST=Used using both a lab devserver and a local devserver
TEST=With Full and Delta payloads
TEST=Tried with payloads from gs://chromeos-image-archive and
gs://chromeos-releases

Change-Id: I0e0064b3301cecb99dc57f7c9eea7ff2bc1d057a
Reviewed-on: https://chromium-review.googlesource.com/544536
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>