Introduce HostInfo object.
HostInfo is a new object that will be the single source of truth of host
information for autoserv and tests as stored in the scheduling
infrastructure.
This will replace all direct access to AFE and other objects that cache
this information for autoserv and the tests.

BUG=chromium:678430
TEST=(new) unittests.

Change-Id: If4f6f730ef456cfb05fa11ba1ade1e7d3565bed4
Reviewed-on: https://chromium-review.googlesource.com/425329
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>