Update cheets_GTS to support version 4.1+.
A recent GTS update fundamentally changes how it works. It enforces its new format on M and N devices breaking
our current workflow. This updates it so that we parse the output correctly from
the new versions.

This will work with waivers. The current waivers don't work because the
tests have been refactored to different packages.

This assumes that java 1.8 is in the lab.

BUG=b:34176927, chromium:683640
TEST=test_that <ip> cheets_GTS.4.1_r1.GtsPlacementTestCases

Change-Id: I5deb69229a0af189b0320e73ee9fc970e4b46728
Reviewed-on: https://chromium-review.googlesource.com/426348
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>