server: Inject a CachingHostInfoStore into Host objects.
This perpetuates the original sin of CL:349990, additionally providing a
CachingHostInfoStore with the frontend.Host object (afe_host).
This is a hack to allow us to easily replace uses of afe_host with
HostInfo. We currently punt on the question of the correct ownership of
HostInfo.

BUG=chromium:678430
TEST=None.

Change-Id: I2586b73d58d9679c032a5f48b357f7d343ea2614
Reviewed-on: https://chromium-review.googlesource.com/426921
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>