Rework adb_host/teststation_host closing.
The process of closing a host calls stop_loggers AFTER the ssh
connection has been disconnected. But for adb_host it grabs the
logs from the remote test station which opens a new ssh connection.

This change ensures that last SSH connection is properly cleaned up.

Lastly, removes the teststation optimization of not calling close, if
its already closed as it will be reopened by each DUT that is part of
the testbed.

BUG=chromium:679024
TEST=local testbed dummy test run and verified no stray ssh connections
remained afterwards.

Change-Id: Ia584813e49061f7f08300965027ac97fc872b44b
Reviewed-on: https://chromium-review.googlesource.com/427361
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>