firmware_ECBootTime: Prevent dropped power on cmd.
firmware_ECBootTime was just checking to see if the EC was up and then
immediately sending the EC command to power on the system.  However,
it's very possible that the console task isn't quite ready at that point
in time and the power on command may be dropped.  This will result in a
faulty test failure.

This commit changes the test such that it waits for the string printed
by the console task indicating that it's ready. This string is the
following:

  "Console is enabled"

This should be much more robust than just searching for "Inits done".

BUG=chrome-os-partner:59923=
BRANCH=None
TEST=Run firmware_ECBootTime against chell and verify it passes.
TEST=Change the queried string to one that is not printed and verify
that the test fails.

Change-Id: I0aa0132ff95272b909ef2951cc31a29b2fa04f8e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/428184
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>