Stability related fix for YouTubeHTML5 on Elm/Oak.
YouTubeHTML5 test was failing on a couple platforms because a
larger version of the stream was getting cached before switching
to a lower stream. As a result, the lower stream wouldn't play.
I have changed the logic of the test and related utility so that
the sequence is more predictable.

BUG=None
TEST=Tested locally on an Elm device.

Change-Id: If1efce2500807c5f89e55d70c7d65f1099d6cce9
Reviewed-on: https://chromium-review.googlesource.com/431420
Commit-Ready: Vinayak Suley <vsuley@chromium.org>
Tested-by: Vinayak Suley <vsuley@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>