gs_offloader: Stop sending emails on offloading failures.
Better monarch based reporting is in place. A list of failed jobs is now
dropped on the server for debugging purposes.

This CL also removes the unittests around email reporting, since the
logic related to deciding when an email report should be sent is now
gone.

BUG=chromium:686904
TEST=None

Change-Id: I2db1463ac72d9c51e598345f02871aba65de744e
Reviewed-on: https://chromium-review.googlesource.com/435103
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>