[autotest] Don't run verify on servo_host when run via test_that
Users of test_that against lab duts will hit failures due to the
servo_repair _UpdateVerifierverification process requiring access
to the devservers in the lab which is not easily do-able due
to subnet restrictions.

Therefore these checks will be skipped when not run in a lab
environment.

BUG=chromium:675284
TEST=test_that chromeos4-row1-rack10-host15.cros platform_SuspendResumeTiming

Change-Id: I971b3ba89208cc64bf4178362b6267deade6a7d7
Reviewed-on: https://chromium-review.googlesource.com/435498
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>